Generate a Dataset with One Categorical Independent Variable and One Continuous Dependent Variable

This function simulates a dataset with multiple groups (levels of one IV), each with a specified mean and a shared standard deviation. The dependent variable is continuous, and the standard deviation is constant across groups, emulating homoscedasticity. The number of observations per group is equal, creating a balanced design.

Usage

DatasetOne(n_sample, means, sds)

Arguments

n_sample

Number of observations per group.

means

A numeric vector of group means.

sds

A single numeric value representing the shared standard deviation.

Value

A data.frame with columns iv (group identifier) and dv (dependent variable).

Examples

set.seed(2025)
DatasetOne(n_sample = 30, means = c(120, 135, 115), sds = 20)
#>      Ind       Dep
#> 1  Gr._1 132.41513
#> 2  Gr._1 120.71283
#> 3  Gr._1 135.46309
#> 4  Gr._1 145.44978
#> 5  Gr._1 127.41951
#> 6  Gr._1 116.74291
#> 7  Gr._1 127.94224
#> 8  Gr._1 118.40021
#> 9  Gr._1 113.10070
#> 10 Gr._1 134.04303
#> 11 Gr._1 112.08607
#> 12 Gr._1  84.89892
#> 13 Gr._1 111.58072
#> 14 Gr._1 135.29819
#> 15 Gr._1 141.32324
#> 16 Gr._1 116.12607
#> 17 Gr._1 118.95706
#> 18 Gr._1 128.35342
#> 19 Gr._1  93.74612
#> 20 Gr._1 168.58038
#> 21 Gr._1 115.10859
#> 22 Gr._1 134.67571
#> 23 Gr._1 177.05563
#> 24 Gr._1 133.43051
#> 25 Gr._1 117.68301
#> 26 Gr._1  93.97296
#> 27 Gr._1 101.64399
#> 28 Gr._1 121.01432
#> 29 Gr._1 112.39734
#> 30 Gr._1 134.45250
#> 31 Gr._2 114.48485
#> 32 Gr._2 171.39741
#> 33 Gr._2 109.59165
#> 34 Gr._2 146.77632
#> 35 Gr._2 122.31493
#> 36 Gr._2 109.68369
#> 37 Gr._2 140.67498
#> 38 Gr._2 105.36478
#> 39 Gr._2 153.23759
#> 40 Gr._2 154.42938
#> 41 Gr._2 160.46254
#> 42 Gr._2 164.24537
#> 43 Gr._2 122.77834
#> 44 Gr._2 127.78894
#> 45 Gr._2 104.41696
#> 46 Gr._2 157.66528
#> 47 Gr._2 126.18522
#> 48 Gr._2 137.59210
#> 49 Gr._2 152.42465
#> 50 Gr._2 145.11029
#> 51 Gr._2 142.41785
#> 52 Gr._2 110.43981
#> 53 Gr._2 131.75605
#> 54 Gr._2 156.76143
#> 55 Gr._2 142.18403
#> 56 Gr._2 156.34382
#> 57 Gr._2 147.56758
#> 58 Gr._2 118.36842
#> 59 Gr._2 121.15229
#> 60 Gr._2 134.10420
#> 61 Gr._3 126.89138
#> 62 Gr._3 127.78202
#> 63 Gr._3 124.58844
#> 64 Gr._3 142.46125
#> 65 Gr._3  92.04632
#> 66 Gr._3  65.98603
#> 67 Gr._3  96.94874
#> 68 Gr._3 107.05137
#> 69 Gr._3 149.33729
#> 70 Gr._3 106.78493
#> 71 Gr._3 113.16433
#> 72 Gr._3 100.06994
#> 73 Gr._3  98.56609
#> 74 Gr._3  68.65961
#> 75 Gr._3 104.29331
#> 76 Gr._3 134.94760
#> 77 Gr._3 104.09564
#> 78 Gr._3 145.19298
#> 79 Gr._3 100.12323
#> 80 Gr._3 101.65122
#> 81 Gr._3  92.58214
#> 82 Gr._3 144.40086
#> 83 Gr._3 119.10513
#> 84 Gr._3  93.77516
#> 85 Gr._3 114.98050
#> 86 Gr._3 114.51106
#> 87 Gr._3 141.08352
#> 88 Gr._3  95.15862
#> 89 Gr._3  94.02833
#> 90 Gr._3 135.17277