Visualizing the Structure of a Linear Model

This function takes a linear model object created with the lm() function and returns tabular summaries describing the model structure. It reports the number and types of variables, levels of factors, and the type of contrast matrices applied to categorical predictors.

Usage

ModelContrastTable(LinearModel, type = "all")

Arguments

LinearModel

A linear model object created using lm().

type

A character string specifying the type of table to return. Options include:

"all" (default): Returns both model-level and variable-level summaries.

"summary": Returns the model-level summary.

"x&y": Returns the variable-level summary (including dependent and independent variables).

"x": Returns a summary of independent variables only.

"y": Returns a summary of the dependent variable only.

Value

A table or a list of tables summarizing the structure of the model, depending on the type argument:

For "all": A list containing both model-level and variable-level summary tables.

For "summary": A data frame summarizing overall model structure.

For "x&y": A data frame with variable names, types, levels, and contrast types.

For "x" or "y": A data frame focused on the specified subset of variables.

Examples

# Creating a linear model with the generative function
DT <- DatasetOne(n_sample = 30, means = c(120, 135, 115), sds = 20)
DT$Ind <- as.factor(DT$Ind)
model <- lm(Dep ~ Ind, data = DT)

# Without specifing the additional parameters (All tables by default)
ModelContrastTable(model)
#> $`Model-Level Summary`
#>                      Metric     Value
#> 1           Total Variables         2
#> 2       Dependent Variables         1
#> 3                Predictors         1
#> 4              Interactions         0
#> 5   Dependent Variable Name       Dep
#> 6  Dependent Variable Types   numeric
#> 7          Dependent Levels       Num
#> 8           Predictor Names       Ind
#> 9           Predictor Types    factor
#> 10         Predictor Levels         3
#> 11    Contrast Matrix Types treatment
#> 
#> $`Variable-Level Summary`
#>   Variable    Type Levels Contrast_Matrix_Type
#> 1      Dep numeric    Num                 <NA>
#> 2      Ind  factor      3            treatment
#> 

# Function with only indipendent variables informations
ModelContrastTable(model, type = "x")
#>   Variable   Type Levels Contrast_Matrix_Type
#> 2      Ind factor      3            treatment